AzStorageServiceTables moved to types file
